refactor: #12 코드 정리
