+++ app/account/index.tsx
@@ -1,4 +1,3 @@
-// 프로필 편집 메인 화면
 import { useRouter } from 'expo-router';
 import styled from 'styled-components/native';
 import { Text, TouchableOpacity } from 'react-native';
@@ -14,37 +13,21 @@ import {
 } from '@/theme/globalStyles';
 import colors from '@/theme/color';
 import { mockUserData } from '@/constants/mockUserData';
-
-const EditButton = ({
-  title,
-  content,
-  onPress,
-  isTextarea = false,
-}: {
-  title: string;
-  content: string;
-  onPress: () => void;
-  isTextarea?: boolean;
-}) => (
-  <EditWrapper>
-    <Title>{title}</Title>
-    {isTextarea ? (
-      <TextareaWrapper onPress={onPress}>
-        <ScrollableTextContainer>
-          <TextareaText>{content}</TextareaText>
-        </ScrollableTextContainer>
-        <CharCount>{content.length}/200</CharCount>
-      </TextareaWrapper>
-    ) : (
-      <Box onPress={onPress}>
-        <BoxText numberOfLines={1}>{content}</BoxText>
-      </Box>
-    )}
-  </EditWrapper>
-);
+import { ProfileEditButton } from '@/components/common/Button/ProfileEditButton';
 
 export default function ProfileEdit() {
   const router = useRouter();
+
+  const editFields = [
+    { title: '카톡 아이디', content: mockUserData.kakaoId, type: 'kakaoId' },
+    { title: 'MBTI', content: mockUserData.mbti, type: 'mbti' },
+    {
+      title: '자기소개',
+      content: mockUserData.intro,
+      type: 'intro',
+      isTextarea: true,
+    },
+  ];
 
   return (
     <Container>
@@ -85,31 +68,17 @@ export default function ProfileEdit() {
         프로필 사진 바꾸기
       </CustomButton>
 
-      <EditButton
-        title="카톡 아이디"
-        content={mockUserData.kakaoId}
-        onPress={() =>
-          router.push({
-            pathname: '/account/edit',
-            params: { type: 'kakaoId' },
-          })
-        }
-      />
-      <EditButton
-        title="MBTI"
-        content={mockUserData.mbti}
-        onPress={() =>
-          router.push({ pathname: '/account/edit', params: { type: 'mbti' } })
-        }
-      />
-      <EditButton
-        title="자기소개"
-        content={mockUserData.intro}
-        onPress={() =>
-          router.push({ pathname: '/account/edit', params: { type: 'intro' } })
-        }
-        isTextarea
-      />
+      {editFields.map(({ title, content, type, isTextarea }) => (
+        <ProfileEditButton
+          key={type}
+          title={title}
+          content={content}
+          isTextarea={isTextarea}
+          onPress={() =>
+            router.push({ pathname: '/account/edit', params: { type } })
+          }
+        />
+      ))}
     </Container>
   );
 }
@@ -123,61 +92,4 @@ const Container = styled.ScrollView`
 const ProfileImageWrapper = styled.View`
   align-items: center;
   margin-top: ${20 * height}px;
-`;
-
-const EditWrapper = styled.View`
-  margin-top: ${20 * height}px;
-`;
-
-const Title = styled.Text`
-  font-size: ${fontSize.sm}px;
-  color: ${colors.text[2]};
-  font-family: ${fonts.Regular};
-  margin-bottom: ${6 * height}px;
-  font-weight: 700;
-`;
-
-const Box = styled.Pressable`
-  width: 100%;
-  border-radius: 8px;
-  padding: ${12 * height}px ${14 * width}px;
-  background-color: transparent;
-  border: 1px solid ${colors.gray[300]};
-`;
-
-const BoxText = styled.Text`
-  font-size: ${fontSize.md}px;
-  color: ${colors.black};
-  font-family: ${fonts.Regular};
-`;
-const TextareaWrapper = styled.Pressable`
-  width: 100%;
-  height: ${76 * height}px;
-  border-radius: 8px;
-  padding: ${12 * height}px ${14 * width}px;
-  border: 1px solid ${colors.gray[300]};
-  position: relative;
-  background-color: transparent;
-`;
-
-const ScrollableTextContainer = styled.ScrollView.attrs({
-  showsVerticalScrollIndicator: false,
-})`
-  max-height: ${36 * height}px;
-`;
-
-const TextareaText = styled.Text`
-  font-size: ${fontSize.md}px;
-  color: ${colors.text[2]};
-  font-family: ${fonts.Regular};
-  line-height: ${lineHeight.l}px;
-`;
-
-const CharCount = styled.Text`
-  position: absolute;
-  bottom: ${12 * height}px;
-  right: ${12 * width}px;
-  font-size: ${fontSize.xs}px;
-  color: ${colors.gray[500]};
-  margin-top: ${12 * height}px;
 `;
